Kernel: Split the ProcFS core file into smaller components
